+++ docs/changes.rst
@@ -25,6 +25,10 @@ Released: not yet
 
 **Incompatible changes:**
 
+* The zhmcclient mock support for Partitions no longer allows to stop a
+  partition when it is in status 'degraded' or 'reservation-error'.
+  That is consistent with the real HMC as described in the HMC API book.
+
 **Deprecations:**
 
 **Bug fixes:**
@@ -43,6 +47,10 @@ Released: not yet
 * Avoided `DeprecationWarning` on Python 3 for invalid escape sequences
   in some places.
 
+* The zhmcclient mock support for various resource classes did not always
+  check for invalid CPC status and for invalid Partition status as
+  described in the HMC API book. It now does.
+  
 **Known issues:**
 
 * See `list of open issues`_.